requirements.txt
flask
mysql-connector-python
python-dotenv
requests
pyserial
boto3
awsiotsdk
paho-mqtt
